+++ src/components/templates/index.ts
@@ -1,0 +1,11 @@
+import UserHomePageLayout from "./UserHomePageLayout";
+import CaptainHomePageLayout from "./CaptainHomePageLayout";
+import RidingCaptainLayout from "./RidingCaptainLayout";
+import RidingUserLayout from "./RidingUserLayout";
+
+export {
+    UserHomePageLayout,
+    CaptainHomePageLayout,
+    RidingCaptainLayout,
+    RidingUserLayout,
+}
